chose: added send form;

+++ src/buttons/acceptTermsButton.js
@@ -15,10 +15,12 @@ module.exports = {
         const form_manager = client.formManager
         const userId = interaction.user.id
         await interaction.deferUpdate()
+        if (form_manager.has(userId)) return await interaction.reply(`Você já aceitou os termos.`)
         try {
             await form_manager.start(interaction)
         } catch (error) {
             form_manager.delete(userId)
+            console.error(error)
             if (error.message) return await interaction.user.send({ content: error.message })
             else return await interaction.user.send({ content: `Você não respondeu o questionário a tempo.` })
         }
